backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
python-multipart>=0.0.9
websockets>=12.0
pytest>=8.0.0
httpx>=0.27.0
